requirements.txt
telethon==1.36.0
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic>=2.10.0
aiofiles==24.1.0
python-multipart==0.0.20
